feat: Add 3D procedural accessories, 3D demon wings, and LoveMachine Fatality-style loader UI

+++ lovemachine/checksum_crc.h
@@ -139,7 +139,7 @@ JustAfew:
 	// The low-order two bits of pb and nBuffer in total control the
 	// upfront work.
 	//
-	nFront = ((unsigned int)pb) & 3;
+	nFront = ((uintptr_t)pb) & 3;
 	nBuffer -= nFront;
 	switch (nFront)
 	{
